[Tests] Move module management code to `conftest.py`
We were duplicating the module unloading code. This moves it to a
shared fixture.

Part of #101.
